payments: relocate shared EPC QR generation code
