+++ apps/payments/invoice.py
@@ -1,5 +1,4 @@
 from decimal import Decimal
-from io import BytesIO
 import logging
 import shutil
 import os.path
@@ -15,8 +14,6 @@ from flask import (
     send_file,
 )
 from flask_login import login_required, current_user
-from markupsafe import Markup
-from segno import helpers
 from sqlalchemy.sql.functions import func
 from wtforms import TextAreaField, SubmitField
 
@@ -24,44 +21,12 @@ from main import external_url, db
 from ..common.receipt import render_pdf
 from models.product import Product, PriceTier
 from models.purchase import Purchase
+from ..common.epc import format_inline_epc_qr
 from ..common.forms import Form
 from . import get_user_payment_or_abort
 from . import payments
 
 logger = logging.getLogger(__name__)
-
-
-def make_epc_qrfile(payment, **kwargs):
-    qrfile = BytesIO()
-    # TODO: this isn't currently used. Need to fetch IBAN from payment.recommended_destination
-    # and name from somewhere - maybe config rather than hard-coding.
-    qr = helpers.make_epc_qr(
-        name="EMF Festivals Ltd",
-        iban=payment.recommended_destination.iban,
-        amount=payment.amount,
-        reference=payment.bankref,
-        encoding=1,
-    )
-    qr.save(qrfile, **kwargs)
-    qrfile.seek(0)
-    return qrfile
-
-
-def qrfile_to_svg(qrfile):
-    return Markup(qrfile.getvalue().decode("utf-8"))
-
-
-def format_inline_epc_qr(payment):
-    qrfile = make_epc_qrfile(
-        payment,
-        kind="svg",
-        svgclass=None,
-        omitsize=True,
-        xmldecl=False,
-        svgns=False,
-        nl=False,
-    )
-    return qrfile_to_svg(qrfile)
 
 
 class InvoiceForm(Form):
